Implement hook to setup mantine and styled components theme.
