+++ graylog2-web-interface/src/theme/GraylogThemeProvider.tsx
@@ -18,18 +18,12 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import { ThemeProvider } from 'styled-components';
 import type { ColorScheme } from '@graylog/sawmill';
-import SawmillSC from '@graylog/sawmill/styled-components';
-import type { MantineTheme } from '@graylog/sawmill/mantine';
-import SawmillMantine from '@graylog/sawmill/mantine';
-import { useMemo } from 'react';
 import { MantineProvider } from '@mantine/core';
 
-import usePluginEntities from 'hooks/usePluginEntities';
-import type { CustomThemesColors } from 'theme/theme-types';
+import useThemes from 'theme/hooks/useThemes';
 
 import ColorSchemeContext from './ColorSchemeContext';
 import { COLOR_SCHEMES } from './constants';
-import usePreferredColorScheme from './hooks/usePreferredColorScheme';
 
 import 'material-symbols/rounded.css';
 
@@ -39,37 +33,8 @@ type Props = {
   userIsLoggedIn: boolean,
 }
 
-const useSCTheme = (
-  changeColorScheme: (newColorScheme: ColorScheme) => void,
-  mantineTheme: MantineTheme,
-) => useMemo(() => {
-  const theme = SawmillSC(mantineTheme);
-
-  return ({
-    ...theme,
-    changeMode: changeColorScheme,
-    mantine: mantineTheme,
-  });
-}, [changeColorScheme, mantineTheme]);
-
-const useMantineTheme = (
-  colorScheme: ColorScheme,
-  useCustomThemeColors: () => ({ data: CustomThemesColors }),
-) => {
-  const { data: customThemeColors } = useCustomThemeColors?.() ?? {};
-
-  return useMemo(() => SawmillMantine({
-    colorScheme,
-    customColors: customThemeColors?.[colorScheme],
-  }), [colorScheme, customThemeColors]);
-};
-
 const GraylogThemeProvider = ({ children, initialThemeModeOverride, userIsLoggedIn }: Props) => {
-  const [colorScheme, changeColorScheme] = usePreferredColorScheme(initialThemeModeOverride, userIsLoggedIn);
-  const themeCustomizer = usePluginEntities('customization.theme.customizer');
-  const useCustomThemeColors = themeCustomizer?.[0]?.hooks.useCustomThemeColors;
-  const mantineTheme = useMantineTheme(colorScheme, useCustomThemeColors);
-  const scTheme = useSCTheme(changeColorScheme, mantineTheme);
+  const { scTheme, mantineTheme, colorScheme } = useThemes(initialThemeModeOverride, userIsLoggedIn);
 
   return (
     <ColorSchemeContext.Provider value={colorScheme}>
